+++ daemon/wui/htmlschedulerconfigitemdelegate.cpp
@@ -46,10 +46,12 @@ QString HtmlSchedulerConfigItemDelegate::text(const QModelIndex &index) const {
                      "<i class=\"icon-play\"></i></span> ");
       else
         text.prepend("<span class=\"label label-danger\" title=\"Remove\">"
-                     "<a href=\"confirm?event=removeConfig&configid="+id
+                     "<a href=\"confirm/do/v1/configs/remove/"+id
+                     +"?confirm_message=remove configuration "+id
                      +"\"><i class=\"icon-trash\"></i></a></span> "
                       "<span class=\"label label-danger\" title=\"Activate\">"
-                      "<a href=\"confirm?event=activateConfig&configid="+id
+                      "<a href=\"confirm/do/v1/configs/activate/"+id
+                     +"?confirm_message=activate configuration "+id
                      +"\"><i class=\"icon-play\"></i></a></span> ");
     }
   }